+++ R/readwrite.R
@@ -89,38 +89,6 @@ ukb_gen_write_plink <- function(x, path, ukb.variables, ukb.id = "eid", na.strin
 
 
 
-#' Writes a PLINK format file for combined exclusions
-#'
-#' @description \strong{Defunct. See help("ukb_defunct")}.
-#
-#' Writes a combined exclusions file including UKB recommended exclusions, heterozygosity exclusions (+/- 3*sd from mean), genetic ethnicity exclusions (based on the UKB genetic ethnic grouping variable, field 1002), and relatedness exclusions (a randomly-selected member of each related pair). For exclusion of individuals from a genetic analysis, the PLINK flag \code{--remove} accepts a space/tab-delimited text file with family IDs in the first column and within-family IDs in the second column (i.e., FID IID), without a header.
-#'
-#' @param path A path to a file.
-#'
-#' @seealso \code{\link{ukb_gen_meta}}, \code{\link{ukb_gen_pcs}} which retrieve variables to be included in a covariate file. \code{\link{ukb_gen_excl_to_na}} to update a phenotype with NAs for samples to-be-excluded based on genetic metadata, and \code{\link{ukb_gen_write_plink}} and \code{\link{ukb_gen_write_bgenie}}
-#'
-#' @importFrom utils write.table
-#' @export
-#' @examples
-#' \dontrun{
-#' # Supply name of a file to write PLINK format combined exclusions
-#' ukb_gen_write_plink_excl("combined_exclusions.txt")
-#' }
-#'
-ukb_gen_write_plink_excl <- function(path) {
-  ukb_defunct()
-  # utils::write.table(
-  #   ukb_meta_excl_plink,
-  #   file = path,
-  #   quote = FALSE,
-  #   row.names = FALSE,
-  #   col.names = FALSE
-  # )
-}
-
-
-
-
 #' Writes a BGENIE format phenotype or covariate file.
 #'
 #' Writes a space-delimited file with a header, missing character set to "-999", and observations (i.e. UKB subject ids) in sample file order. Use this function to write phenotype and covariate files for downstream genetic analysis in \href{https://jmarchini.org/bgenie/}{BGENIE} - the format is the same.