+++ R/utils.R
@@ -1,24 +1,4 @@
 
-# Collects a column from a tidyverse tibble as an atomic vector.
-#
-# A function found in a
-# \href{http://stackoverflow.com/a/24730843/3819656}{Stack Overflow} discussion
-# to retrieve a  tibble column as a vector
-#
-# @param data A tidyverse tibble, tbl_df
-# @param y The column to be collected. Can be a numeric index or the column
-#   name (quoted or not)
-#
-pull <- function(data, y) {
-  data[,
-       if (is.name(substitute(y))) {
-         deparse(substitute(y))
-       } else y,
-       drop = FALSE][[1]]
-}
-
-
-
 # Draws a grid of ggplot figures
 #
 # For arranging multiple ggplot2 plots in a grid. Found in the \href{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}{Cookbook for R}.